backend/requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
pydantic==2.9.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
python-multipart==0.0.9
strawberry-graphql>=0.284,<0.285
